+++ src/test/java/com/github/alexisjehan/mvncheck/ApplicationTest.java
@@ -36,6 +36,7 @@ import com.github.alexisjehan.mvncheck.core.component.build.resolver.BuildResolv
 import com.github.alexisjehan.mvncheck.core.util.GithubUtils;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.AdditionalAnswers;
 import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
@@ -319,7 +320,8 @@ final class ApplicationTest {
 	}
 
 	@Test
-	void testRun() throws IOException {
+	@Deprecated
+	void testRunDeprecated3() throws IOException {
 		final var path1 = Path.of("path1");
 		final var path2 = Path.of("path2");
 		final var path3 = Path.of("path3");
@@ -441,7 +443,8 @@ final class ApplicationTest {
 	}
 
 	@Test
-	void testRunInvalid() {
+	@Deprecated
+	void testRunDeprecated3Invalid() {
 		try (var printStream = new PrintStream(OutputStream.nullOutputStream())) {
 			final var application = new Application(printStream);
 			assertThatNullPointerException().isThrownBy(
@@ -465,6 +468,187 @@ final class ApplicationTest {
 							-1,
 							false,
 							false,
+							false
+					)
+			);
+		}
+	}
+
+	@Test
+	void testRun() throws IOException {
+		final var path1 = Path.of("path1");
+		final var path2 = Path.of("path2");
+		final var path3 = Path.of("path3");
+		final var buildFile1 = new BuildFile(BuildFileType.MAVEN, Path.of("pom.xml"));
+		final var buildFile2 = new BuildFile(BuildFileType.GRADLE_GROOVY, Path.of("build.gradle"));
+		final var buildFile3 = new BuildFile(BuildFileType.GRADLE_KOTLIN, Path.of("build.gradle.kts"));
+		final var artifact = new Artifact<>(
+				MavenArtifactType.DEPENDENCY,
+				new ArtifactIdentifier("foo-group-id", "foo-artifact-id"),
+				"1.0.0",
+				true
+		);
+		final var build1 = new Build(
+				buildFile1,
+				List.of(),
+				List.of(artifact)
+		);
+		final var build2 = new Build(
+				buildFile2,
+				List.of(),
+				List.of(artifact)
+		);
+		Mockito.when(mockedService.findBuildFiles(Mockito.argThat(path1::equals), Mockito.anyInt()))
+				.thenReturn(List.of());
+		Mockito.when(mockedService.findBuildFiles(Mockito.argThat(path2::equals), Mockito.anyInt()))
+				.thenReturn(List.of(buildFile1));
+		Mockito.when(mockedService.findBuildFiles(Mockito.argThat(path3::equals), Mockito.anyInt()))
+				.thenReturn(List.of(buildFile1, buildFile2, buildFile3));
+		Mockito.when(mockedService.filterBuildFiles(Mockito.anyList()))
+				.then(AdditionalAnswers.returnsFirstArg());
+		Mockito.when(mockedService.findBuild(Mockito.argThat(buildFile1::equals)))
+				.thenReturn(build1);
+		Mockito.when(mockedService.findBuild(Mockito.argThat(buildFile2::equals)))
+				.thenReturn(build2);
+		Mockito.when(mockedService.findBuild(Mockito.argThat(buildFile3::equals)))
+				.thenThrow(BuildResolveException.class);
+		Mockito.when(
+						mockedService.findArtifactUpdateVersions(
+								Mockito.argThat(build1::equals),
+								Mockito.anyBoolean(),
+								Mockito.anyBoolean()
+						)
+				)
+				.thenReturn(List.of());
+		Mockito.when(
+						mockedService.findArtifactUpdateVersions(
+								Mockito.argThat(build2::equals),
+								Mockito.anyBoolean(),
+								Mockito.anyBoolean()
+						)
+				)
+				.thenReturn(List.of(new ArtifactUpdateVersion(artifact, "2.0.0")));
+		try (var printStream = new PrintStream(OutputStream.nullOutputStream())) {
+			try (var mockedStaticGithubUtils = Mockito.mockStatic(GithubUtils.class)) {
+				mockedStaticGithubUtils.when(
+								() -> GithubUtils.retrieveOptionalLatestReleaseName(
+										Mockito.notNull(),
+										Mockito.notNull()
+								)
+						)
+						.thenReturn(
+								Optional.of("1.0.0"),
+								Optional.of("2.0.0"),
+								Optional.empty()
+						);
+				try (var mockedStaticApplication = Mockito.mockStatic(Application.class)) {
+					mockedStaticApplication.when(Application::createService)
+							.thenReturn(mockedService);
+					mockedStaticApplication.when(Application::getCurrentVersion)
+							.thenReturn(
+									"1.0.0",
+									"1.0.0",
+									null
+							);
+					final var application = new Application(printStream);
+					assertThatNoException().isThrownBy(
+							application::run
+					);
+					assertThatNoException().isThrownBy(
+							() -> application.run("--" + Application.OPTION_MAX_DEPTH, "0")
+					);
+					assertThatNoException().isThrownBy(
+							() -> application.run("--" + Application.OPTION_HELP)
+					);
+					assertThatNoException().isThrownBy(
+							() -> application.run("--" + Application.OPTION_IGNORE_SNAPSHOTS)
+					);
+					assertThatNoException().isThrownBy(
+							() -> application.run("--" + Application.OPTION_IGNORE_INHERITED)
+					);
+					assertThatNoException().isThrownBy(
+							() -> application.run("--" + Application.OPTION_INCLUDE_OUTPUT)
+					);
+					assertThatNoException().isThrownBy(
+							() -> application.run("--" + Application.OPTION_SHORT)
+					);
+					assertThatNoException().isThrownBy(
+							() -> application.run("--" + Application.OPTION_VERSION)
+					);
+					assertThatNoException().isThrownBy(
+							() -> application.run("directory_not-found")
+					);
+					assertThatNoException().isThrownBy(
+							() -> application.run(path1.toString())
+					);
+					assertThatNoException().isThrownBy(
+							() -> application.run(path2.toString())
+					);
+					assertThatNoException().isThrownBy(
+							() -> application.run(path3.toString())
+					);
+					assertThatNoException().isThrownBy(
+							() -> application.run(
+									path1,
+									0,
+									true,
+									true,
+									false,
+									true
+							)
+					);
+					assertThatNoException().isThrownBy(
+							() -> application.run(
+									path2,
+									0,
+									true,
+									true,
+									false,
+									true
+							)
+					);
+					assertThatNoException().isThrownBy(
+							() -> application.run(
+									path3,
+									0,
+									true,
+									true,
+									false,
+									true
+							)
+					);
+				}
+			}
+		}
+	}
+
+	@Test
+	void testRunInvalid() {
+		try (var printStream = new PrintStream(OutputStream.nullOutputStream())) {
+			final var application = new Application(printStream);
+			assertThatNullPointerException().isThrownBy(
+					() -> application.run((String[]) null)
+			);
+			assertThatNullPointerException().isThrownBy(
+					() -> application.run((String) null)
+			);
+			assertThatNullPointerException().isThrownBy(
+					() -> application.run(
+							null,
+							0,
+							false,
+							false,
+							true,
+							false
+					)
+			);
+			assertThatIllegalArgumentException().isThrownBy(
+					() -> application.run(
+							Path.of("path"),
+							-1,
+							false,
+							false,
+							true,
 							false
 					)
 			);